削除処理をAjax化

+++ app/Controller/PostsController.php
@@ -37,12 +37,19 @@ class PostsController extends AppController{
 	}
 	public function delete($id) {
 		if ($this->request->is('get')) {
-		throw new MethodNotAllowedException();
+			throw new MethodNotAllowedException();
 		}
-		if ($this->Post->delete($id)) {
-			$this->Session->setFlash('Deleted!');
-			$this->redirect(array('action'=>'index'));
+		if($this->request->is('ajax')) {
+    		if($this->Post->delete($id)) {
+        	$this->autoRender = false;
+        	$this->autoLayout = false;
+        	$response = array('id' => $id);
+        	$this->header('Content-Type: application/json');
+        	echo json_encode($response);
+        	exit();
+    	}
+    	$this->redirect(array('action'=>'index'));
 		}
-	}
+	}	
 }
 ?>
